stylegan_code_finder/requirements.txt
kornia
numpy
opencv-python-headless
Pillow
PyYAML
requests
scikit-image
scipy
tensorboard-pytorch
torch==1.9.0+cu111
torchvision==0.10.0+cu111
torchaudio===0.9.0
tqdm
wandb
ipython
tensorboard
statsmodels
pytorch-fid
imgaug
scikit-learn==0.24.2
flask
pycocotools
csaps
tabulate
pandas

# TransUNet
ml-collections
medpy
h5py
